requirements.txt
accelerate==1.13.0
aiohappyeyeballs==2.6.1
aiohttp==3.13.5
aiosignal==1.4.0
altair==6.0.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
attrs==26.1.0
autocommand==2.2.2
backports.tarfile==1.2.0
blinker==1.9.0
cachetools==7.0.5
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.6
click==8.3.1
cryptography==46.0.6
datasets==4.8.4
dill==0.4.1
distro==1.9.0
duckdb==1.5.1
evaluate==0.4.6
filelock==3.25.2
frozenlist==1.8.0
fsspec==2026.2.0
gitdb==4.0.12
gitpython==3.1.46
h11==0.16.0
hf-xet==1.4.3
httpcore==1.0.9
httpx-sse==0.4.3
httpx==0.28.1
huggingface-hub==0.36.2
idna==3.11
importlib-metadata==8.7.1
iniconfig==2.3.0
jaraco-context==6.1.0
jaraco-functools==4.4.0
jaraco.text==4.0.0
jinja2==3.1.6
jiter==0.13.0
joblib==1.5.3
jsonpatch==1.33
jsonpointer==3.1.1
jsonschema-specifications==2025.9.1
jsonschema==4.26.0
langchain-core==1.2.23
langgraph-checkpoint==4.0.1
langgraph-prebuilt==1.0.8
langgraph-sdk==0.3.12
langgraph==1.1.4
langsmith==0.7.23
markdown-it-py==4.0.0
markupsafe==3.0.3
mcp==1.26.0
mdurl==0.1.2
more-itertools==10.8.0
mpmath==1.3.0
multidict==6.7.1
multiprocess==0.70.19
narwhals==2.18.1
networkx==3.6.1
numpy==2.4.4
openai==2.30.0
orjson==3.11.8
ormsgpack==1.12.2
packaging==26.0
pandas==3.0.2
pillow==12.1.1
platformdirs==4.4.0
pluggy==1.6.0
propcache==0.4.1
protobuf==7.34.1
psutil==7.2.2
pyarrow==23.0.1
pycparser==3.0
pydantic-core==2.41.5
pydantic-settings==2.13.1
pydantic==2.12.5
pydeck==0.9.1
pygments==2.20.0
pyjwt==2.12.1
pytest==9.0.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.22
pyyaml==6.0.3
referencing==0.37.0
regex==2026.3.32
requests-toolbelt==1.0.0
requests==2.33.1
rich==14.3.3
rpds-py==0.30.0
safetensors==0.7.0
scikit-learn==1.8.0
scipy==1.17.1
sentence-transformers==5.3.0
setfit==1.1.3
setuptools==81.0.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.3
sniffio==1.3.1
socksio==1.0.0
sse-starlette==3.3.4
starlette==1.0.0
streamlit==1.56.0
sympy==1.14.0
tenacity==9.1.4
threadpoolctl==3.6.0
tokenizers==0.21.4
toml==0.10.2
tomli==2.4.0
torch==2.11.0
tornado==6.5.5
tqdm==4.67.3
transformers==4.51.3
typer==0.24.1
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
uuid-utils==0.14.1
uvicorn==0.42.0
xxhash==3.6.0
yarl==1.23.0
zipp==3.23.0
zstandard==0.25.0
